moved the wave-drivers so that we don't get bit by relative paths being different
