+++ app/wave-5-game.rb
@@ -1,4 +1,4 @@
-require_relative 'lib/adagrams'
+require_relative '../lib/adagrams'
 
 def display_welcome_message
   puts "Welcome to Adagrams!"
@@ -13,7 +13,12 @@ def display_game_instructions
   puts "Please input your submission for the longest anagram you can come up with"
 end
 
-def display_needs_valid_input_message
+def display_not_in_dictionary_message
+  puts "You entered in a word that is not contained in the English dictionary"
+  display_game_instructions
+end
+
+def display_not_in_letter_bank_message
   puts "You entered in a word that contains characters not in the letter bank"
   display_game_instructions
 end
@@ -27,8 +32,13 @@ def display_retry_instructions
   puts "Enter y to replay"
 end
 
-def display_goodbye_message
+def display_highest_score(high_score_hash)
   puts "Thanks for playing Adagrams!"
+  puts "The highest score from this game was #{high_score_hash[:word]}, which was worth #{high_score_hash[:score]} points"
+end
+
+def display_goodbye_message
+  puts "Goodbye!"
 end
 
 def get_user_input
@@ -36,33 +46,42 @@ def get_user_input
 end
 
 def run_game
+  played_words = []
+  
   display_welcome_message
-
+  
   should_continue = true
-
+  
   while should_continue
     puts "Let's draw 10 letters from the letter pool..."
-
+    
     letter_bank = draw_letters
     display_drawn_letters(letter_bank)
-
+    
     display_game_instructions
-
+    
     user_input_word = get_user_input
-
-    while ( !(uses_available_letters?(user_input_word, letter_bank)) )
-      display_needs_valid_input_message
+    
+    while ( (!(uses_available_letters?(user_input_word, letter_bank)) ) || (!(is_in_english_dict?(user_input_word))) )
+      
+      if !(uses_available_letters?(user_input_word, letter_bank))
+        display_not_in_letter_bank_message
+      elsif !(is_in_english_dict?(user_input_word))
+        display_not_in_dictionary_message
+      end
       user_input_word = get_user_input
     end
-
+    
     score = score_word(user_input_word)
-
+    played_words << user_input_word
+    
     display_score(score)
-
+    
     display_retry_instructions
     should_continue = get_user_input == "y"
   end
-
+  
+  display_highest_score(highest_score_from(played_words))
   display_goodbye_message
 end
